# [OpenCL] specialize convolutionK