Pass some of the properties as constants to enable more aggressive
constant propagation, strength reduction, loop removal, and other
optimizations.

This helps targets that use the non-local memory using convolution
since not having a local scratch pad memory.

Some numbers:

Intel SDK:
inception_v1: From 1.36 to 1.27 (6% drop)
with convolutionW total from 917 ms to 587 ms (36% drop)

mobilenetv2:  From 0.72 to 0.69 (4% drop)
with convolutionW total from 333 ms to 204 ms (39% drop)

resnet50:     From 2.46 to 2.38 (3% drop)
with convolutionW total from 2302 to 1687 ms (27% drop)

pocl:
inception_v1: From 0.57 s to 0.50 s (12% drop)
with convolutionW total from 480 ms to 437 ms (9% drop)

mobilenetv2:  From 0.22 s to 0.17 s (22% drop)
with convolutionW total from 207 ms to 158 ms (23% drop)

resnet50:    From 1.47 s to 1.36 s (7% drop)
with convolutionW total from 1369 ms to 1323 ms (5% drop)

All of these are “best out of 10 runs” numbers on an Intel i7-7500U
CPU OpenCL device.

Test Plan:

Tested with pocl and Intel OpenCL implementations on an Intel i7 CPU.