@pjaaskel 2020-01-30
> Have you tried to see if this PR results in any performance boost on GPUs? After all, the primary use-case for the OpenCL backend are GPUs. If it does not improve the GPU performance it is not clear if the introduced complexity is worth it.

Yes, I've tested on a GPU which benefits from this, but it's not a public one. This kernel is selected instead of conv_forward_mem in case local memory usage is not beneficial for the device at hand, so for such GPUs this patch should be beneficial. For conv_forward_mem there is already a similar specialization mechanism in place.

Also to me OpenCL is about more than GPUs; heterogeneous computing across "device categories", even though it is seen by most of the community as simply a GPU offload API. This improves (multicore) CPU performance via easier work-group SIMDization etc. and likely helps DSP targets / co-processors without scratch pad memories too.