@opti-mix 2020-01-30
@pjaaskel I'll try your PR locally to check the performance on my GPU.