@pjaaskel 2020-01-31
> @pjaaskel I'll try your PR locally to check the performance on my GPU.

Likely you won't see any difference since your GPU probably has proper local memory and thus the other conv_forward_mem gets used instead.