@pjaaskel 2020-02-01
I separated the specialized conv to another file, which made the implementation cleaner. I also cleaned up the implementation a bit while at it (cl_uint32_t to unsigned, const and restrict qualifiers...).