@opti-mix 2020-02-06
@pjaaskel A very minor cosmetic request: Could you rename the `specialized_no_local_mem_conv.cl` file into `kernels_specialized_no_local_mem_conv.cl` to follow the naming convention of other `*.cl` files?